+++ CHANGES.rst
@@ -10,11 +10,12 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- Make ZServer optional
 
 Bug fixes:
 
-- *add item here*
+- More fixes for Python 2 / 3 compatibility
+  [pbauer]
 
 
 1.2.2 (2018-02-11)